# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handlers.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      password:
        type: string
    required:
    - new_password
    - password
    type: object
...
  handlers.ForgetPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  handlers.ResetPasswordRequest:
    properties:
      access_key:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - access_key
    - email
    - password
    type: object
...
  handlers.SignInRequest:
    properties:
      password:
        description: 密碼
        example: password123
        type: string
      username:
        description: 使用者名稱
        example: testuser
        type: string
    required:
    - password
    - username
    type: object
  handlers.SignUpRequest:
    properties:
      email:
        description: 電子郵件
        example: test@example.com
        type: string
      invitation_code:
        description: 邀請碼
        example: sixpens
        type: string
      password:
        description: 密碼
        example: password123
        type: string
      username:
        description: 使用者名稱
        example: testuser
        type: string
    required:
    - email
    - invitation_code
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@higgstv.com
    name: API Support
    url: http://www.higgstv.com
  description: Session-based authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HiggsTV API Server
  version: "1.0"
paths:
...
  /apis/change_password:
    post:
      consumes:
      - application/json
      description: 修改密碼（需登入）
      parameters:
      - description: 變更密碼請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 缺少欄位" example({"state":1,"code":0})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 變更密碼
      tags:
      - 認證
...
  /apis/forget_password:
    post:
      consumes:
      - application/json
      description: 請求寄送重設密碼郵件。無論 Email 是否存在，都回成功（安全設計）
      parameters:
      - description: 忘記密碼請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 缺少欄位" example({"state":1,"code":0})
          schema:
            additionalProperties: true
            type: object
      summary: 忘記密碼
      tags:
      - 認證
...
  /apis/reset_password:
    post:
      consumes:
      - application/json
      description: 使用 access_key 重設密碼（不需要登入）
      parameters:
      - description: 重設密碼請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 缺少欄位" example({"state":1,"code":0})
          schema:
            additionalProperties: true
            type: object
      summary: 重設密碼
      tags:
      - 認證
...
  /apis/signin:
    post:
      consumes:
      - application/json
      description: 使用者登入，成功後會設定 Session Cookie
      parameters:
      - description: 登入請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 失敗回應" example({"state":0,"ret":false})
          schema:
            additionalProperties: true
            type: object
      summary: 登入
      tags:
      - 認證
  /apis/signout:
    get:
      description: 使用者登出，清除 Session。可選的 redirect 參數會執行 HTTP redirect
      parameters:
      - description: 重定向 URL（選填）
        in: query
        name: redirect
        type: string
      responses:
        "200":
          description: 成功回應（無 redirect 時）" example({"state":0})
          schema:
            additionalProperties: true
            type: object
        "302":
          description: 重定向到指定 URL 或首頁
      summary: 登出
      tags:
      - 認證
  /apis/signup:
    post:
      consumes:
      - application/json
      description: 新使用者註冊，需要有效的邀請碼
      parameters:
      - description: 註冊請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 失敗回應" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      summary: 註冊
      tags:
      - 認證
...
swagger: "2.0"